Use requirements.in for install_requires
Parsing setup.cfg (or running setup.py) to extract dependencies is
problematic and often requires workarounds like `-e .` dependencies.
Instead, use requirements.in and load it from setup.py as for
tests_require.

Also use requirements.in to install test dependencies for tox (a big
motivation for the change) and remove constraints.txt which was a
remnant from a previous approach to dependency management.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>