Drop support for Python 3.4
`import pkg_resources` (and therefore `import setuputils`) fails with:

    RuntimeError: Python 3.5 or later is required

This complicates testing with tox, which would require using an old
version of setuputils (in pyproject.toml requires and possibly
elsewhere) on Python 3.4.

Considering that 3.4 hasn't been supported upstream for quite a while,
it's not currently worth the effort to maintain support.  Drop it.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>