Remove version-constrained requirements
These requirements files are intended for applications (rather than
libraries) and would need to be frequently updated to address security
issues and bug fixes.  The maintenance effort is not justified for the
template, and notifications from GitHub et al. for outdated dependencies
due to the versions in these files is annoying.  Remove them.  Keep the
files as references.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>