tox.ini: comment for --cov with package filename
It appears that `--cov={envsitepackagesdir}/packagename` does not work
if packagename is a file (`packagename.py`) rather than a directory
(`packagename/__init__.py`).  It causes:

    warning: Module /.../packagename.py was never imported. (module-not-imported)
    Coverage.py warning: No data was collected. (no-data-collected)
    /.../pytest_cov/plugin.py:271: PytestWarning: Failed to generate report: No data to report.

e.g. https://github.com/kevinoid/ytdl2rss/runs/2485750355.

For packages where the package is a file, use --cov=packagename (and
hope it resolves to the one in the tox venv - which it currently does
for ytdl2rss).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>