isort: include all standard library modules
Recognize modules which are in the Python 2 standard library, but not
Python 3, as standard library modules.  This supports formatting for
projects which still support Python 2 (or have remaining imports for
Python 2 in try/except blocks).

Projects not supporting Python 2 may wish to use "py3" (the default) to
avoid the (small) risk of false positives when using a third-party
module with the same name as the Python 2 standard library module.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>