test(tox_pip_install): use subprocess instead of os
Use subprocess.run instead of os.spawn and os.exec, as recommended by
the Python API docs.

Importantly: On Windows, os.exec creates a child process then exits
without waiting for the child to finish.  This causes the caller to
think installation has finished when it has not.  Using subprocess.run
avoids the issue by always waiting for the child to complete (although
it can be less efficient on Unix-like systems, since the parent process
is not replaced, but sticks around until both finish).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>