tox: remove install_command bug workaround
Remove workaround for https://bugs.debian.org/962654, which was fixed by
pypy3 7.3.4+dfsg-1.

Note: The workaround caused problems for PyPy on Windows due to use of
os.exec and is likely to cause others.  The risks of the workaround
likely outweigh the benefits.  If this bug needs to be worked around
again, consider less risky fixes.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>