.gitlab-ci: switch from alpine to slim images
The regex module required by black does not provide a wheel for Alpine
(musl) and installing tools to build from source is not ideal.
Apparently this is a common issue, along with a few other performance
and compatibility issues, and the current consensus is to avoid Alpine
for Python for now.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>