Use the MIT License for projects
Since GNU and OSI recommend against CC0 for software (due to the
explicit patent disclaimer) and the MIT license is the most popular on
PyPI and it matches my intent for my most common projects, use it
instead.

Keep the template CC0 to make it easy for anyone to use for any purpose
without the need to keep any of its licensing information in software
based on the template.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>