test: fix EncodingWarning for Python < 3.10
EncodingWarning was introduced in Python 3.10.  When running on Python <
3.10, ignore::EncodingWarning produces

> AttributeError: module 'builtins' has no attribute 'EncodingWarning'

See https://github.com/kevinoid/python-project-template/runs/6510589679

To avoid this, enable and ignore EncodingWarning only for the py310 tox
environment.  Note that python -W options match after pytest
filterwarnings, so ignore::EncodingWarning matches after
error:::packagename which causes EncodingWarning to throw for code in
this project and be ignored for other code, as we want.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>