drop support for Python 3.9
Python 3.9 has not been supported upstream since 2025-10-31:
https://devguide.python.org/versions/

Python 3.9 is not supported by pytest 9.  This causes CI failures
because pytest8 does not support the strict config option:

> ERROR: Unknown config option: strict

There's currently no reason for this project template to support 3.9
when the rest of the ecosystem does not.  Drop it.

BREAKING CHANGE: Drop support for Python 3.9.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>